Add servlet filter to avoid lazy exeption, small changes in controller, fix abstract entity

+++ src/main/java/by/intexsoft/backend/model/AbstractEntity.java
@@ -7,7 +7,7 @@ import javax.persistence.*;
  * Abstract entity, including same fields from subclasses
  */
 @MappedSuperclass
-public class AbstractEntity extends AbstractPersistable<Integer> {
+public class AbstractEntity extends AbstractPersistable<Long> {
 
     private static final long serialVersionUID = 7219127246357342388L;
 }
